# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RISKOS - AI Payment Risk Engine",
    "description": "Autonomous payment fraud detection, AI investigation, entity graph, and counterfactual simulation platform.",
    "version": "1.0.0"
  },
  "paths": {
    "/api/auth/login": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Login",
        "description": "Authenticate an analyst/admin user and return a JWT Bearer token.\nDefault credentials for demo:\n- Email: analyst@riskos.ai | Password: analyst_demo_secret_2026\n- Email: priya@riskos.ai   | Password: analyst_demo_secret_2026\n- Email: admin@riskos.ai   | Password: admin_demo_secret_2026",
        "operationId": "login_api_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/me": {
      "get": {
        "tags": [
          "Authentication"
        ],
        "summary": "Get Current User Profile",
        "description": "Return currently authenticated analyst profile (protected route).",
        "operationId": "get_current_user_profile_api_auth_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/counterfactual": {
      "post": {
        "tags": [
          "Counterfactual"
        ],
        "summary": "Run Counterfactual",
        "description": "Simulate outcomes of allow/verify/hold/block for a transaction.",
        "operationId": "run_counterfactual_api_counterfactual_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "type": "object",
                "title": "Body"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CounterfactualResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/campaigns/{campaign_id}": {
      "get": {
        "tags": [
          "Campaign Detection"
        ],
        "summary": "Get Campaign Detail",
        "description": "Fetch full campaign detail including entities, full timeline events, and suppressor proof.",
        "operationId": "get_campaign_detail_api_campaigns__campaign_id__get",
        "parameters": [
          {
            "name": "campaign_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Campaign Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/campaigns/{campaign_id}/contain": {
      "post": {
        "tags": [
          "Campaign Detection"
        ],
        "summary": "Contain Campaign Endpoint",
        "description": "Execute containment policy across all entities attributed to this campaign.",
        "operationId": "contain_campaign_endpoint_api_campaigns__campaign_id__contain_post",
        "parameters": [
          {
            "name": "campaign_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Campaign Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContainmentRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/campaigns/{campaign_id}/counterfactual": {
      "post": {
        "tags": [
          "Campaign Detection"
        ],
        "summary": "Get Campaign Counterfactual Endpoint",
        "description": "Compute campaign containment counterfactual comparison (Status Quo vs Challenge vs Contain).",
        "operationId": "get_campaign_counterfactual_endpoint_api_campaigns__campaign_id__counterfactual_post",
        "parameters": [
          {
            "name": "campaign_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Campaign Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CampaignPolicy": {
        "type": "string",
        "enum": [
          "allow",
          "challenge",
          "contain"
        ],
        "title": "CampaignPolicy"
      },
      "ContainmentRequest": {
        "properties": {
          "policy": {
            "$ref": "#/components/schemas/CampaignPolicy",
            "description": "Policy to apply: contain or challenge",
            "default": "contain"
          },
          "analyst_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Analyst Name",
            "default": "Demo Analyst"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note",
            "default": "Coordinated botnet containment executed from RISKOS console."
          }
        },
        "type": "object",
        "title": "ContainmentRequest"
      },
      "CounterfactualOption": {
        "properties": {
          "action": {
            "type": "string",
            "title": "Action"
          },
          "estimated_fraud_loss_paise": {
            "type": "integer",
            "title": "Estimated Fraud Loss Paise"
          },
          "estimated_fraud_loss_inr": {
            "type": "number",
            "title": "Estimated Fraud Loss Inr"
          },
          "friction_cost_paise": {
            "type": "integer",
            "title": "Friction Cost Paise"
          },
          "friction_cost_inr": {
            "type": "number",
            "title": "Friction Cost Inr"
          },
          "net_expected_value_paise": {
            "type": "integer",
            "title": "Net Expected Value Paise"
          },
          "recommended": {
            "type": "boolean",
            "title": "Recommended"
          }
        },
        "type": "object",
        "required": [
          "action",
          "estimated_fraud_loss_paise",
          "estimated_fraud_loss_inr",
          "friction_cost_paise",
          "friction_cost_inr",
          "net_expected_value_paise",
          "recommended"
        ],
        "title": "CounterfactualOption"
      },
      "CounterfactualResponse": {
        "properties": {
          "transaction_id": {
            "type": "string",
            "title": "Transaction Id"
          },
          "risk_score": {
            "type": "number",
            "title": "Risk Score"
          },
          "options": {
            "items": {
              "$ref": "#/components/schemas/CounterfactualOption"
            },
            "type": "array",
            "title": "Options"
          },
          "recommended_action": {
            "type": "string",
            "title": "Recommended Action"
          }
        },
        "type": "object",
        "required": [
          "transaction_id",
          "risk_score",
          "options",
          "recommended_action"
        ],
        "title": "CounterfactualResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LoginRequest": {
        "properties": {
          "email": {
            "type": "string",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "LoginRequest"
      },
...
      "TokenResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          },
          "expires_in_minutes": {
            "type": "integer",
            "title": "Expires In Minutes"
          },
          "user": {
            "$ref": "#/components/schemas/UserInfo"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "expires_in_minutes",
          "user"
        ],
        "title": "TokenResponse"
      },
...
      "UserInfo": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "role": {
            "type": "string",
            "title": "Role"
          }
        },
        "type": "object",
        "required": [
          "name",
          "email",
          "role"
        ],
        "title": "UserInfo"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}